[BUGFIX] Profile or password editing may fail
Problem: When a form contains rsa-encryption input fields, on form
submission, the form is intercepted, the fields are encrypted, and
the form is re-submitted. However, the submit input field is not
submitted. When there is no preview form, the update fails.

Solution: Add a hidden input field with same name as the submit input
element.

Resolves: #76119
Related: #76120
Change-Id: I0f5f07b9ab02a33b36a05d654ab22e81e93d30be
Reviewed-on: https://review.typo3.org/55608
Reviewed-by: Stanislas Rolland <typo3@sjbr.ca>
Tested-by: Stanislas Rolland <typo3@sjbr.ca>