[BUGFIX] Misleading frontend output when deleting images in edit mode
If a frontend user is trying to delete an existing profile image
using the trashcan symbol shown next to the image file name
the page and form is reloaded and the image is removed from file
system but still not removed from user record in database.
This issue arises with TYPO3 7 LTS, but not with TYPO3 8 LTS.
With TYPO3 8 LTS, the file and the reference are deleted as soon as
the trash can is hit, so that when the form is reloaded the file is
not shown anymore, even though the whole form is not yet saved.
This solution will be applied to TYPO3 7 LTS as well.

Resolves: #84569
Change-Id: Ibe52d0c1930c069c1cdb5c8e923fc636cf0717e0
Reviewed-on: https://review.typo3.org/58679
Reviewed-by: Stanislas Rolland <typo3@sjbr.ca>
Tested-by: Stanislas Rolland <typo3@sjbr.ca>